movies_series_api/requirements.txt
tox
Django
Pillow
django-filter
drf-flex-fields
drf-access-policy
djangorestframework
django-versatileimagefield